\name{hypergea-package}
\alias{hypergea-package}
\alias{hypergea}
\docType{package}
\title{
Hypergeometric tests
}
\description{
The package hypergea provides parallelized methods to perform exact hypergeometric tests on IxJ and 2x2x2 contingency tables
}
\details{
%\tabular{ll}{
%Package: \tab hypergea\cr
%Type: \tab Package\cr
%Version: \tab 1.0\cr
%Date: \tab 2014-06-16\cr
%License: \tab GPL (>= 2)\cr
%}
The package enables users to perform exact tests for IxJ and 2x2x2 contingency tables.
}
\author{
Markus Boenn
}
\keyword{ package }
