\name{hypergea-package}
\alias{hypergea-package}
\alias{hypergea}
\docType{package}
\title{
Hypergeometric tests
}
\description{
The package hypergea provides parallelized methods to perform exact hypergeometric tests on \eqn{I\times J}{IxJ} and \eqn{2\times2\times2}{2x2x2} contingency tables.
}
\details{
The package enables users to perform exact tests for \eqn{I\times J}{IxJ} and \eqn{2\times2\times2}{2x2x2} contingency tables. For this, a C code implementation is used with a parallelized loop (OpenMP) to allow exact test even for large values in the contingency tables in little time.
}
\author{
Markus Boenn
}
\keyword{ package }
\keyword{ OpenMP }
